move sw and update resources
