+++ client/src/index.html
@@ -2,7 +2,7 @@
 <html>
   <head>
     <meta charset="UTF-8">
-    <!-- https://tomitm.github.io/appmanifest/ -->
+    <link rel="icon" href="assets/favicon.ico">
     <title>Mapitout</title>
     <meta name="mobile-web-app-capable" content="yes">
     <meta name="apple-mobile-web-app-capable" content="yes">
